refactor: address page-specific component issues and improve UX
- Refactor Dashboard.tsx with Suspense and lazy-loaded sub-components.
- Add loading/error states to Blog and Resources features.
- Decouple Contact presentation from its logic.
- Optimize useParams usage in dynamic routes to reduce prop-drilling.
